refactor(router): split routes/handoff.rs (q1 wave 4/5)
Extract spawn + handoff route builders and 13 handlers from routes/mod.rs
into routes/handoff.rs (~1600 LOC). Pure mechanical split plus one
namespace-resolution fix.

## Byte-equivalence proof

Drift checker (syn-based extractor hashing every fn body) reports ONE
function body mutation, by design:

  handoff_protected_routes:
    super::handoff::MAX_BLOB_SIZE_BYTES → handoff::MAX_BLOB_SIZE_BYTES

Originally at routes.rs top-level, `super::handoff` resolved to
`crate::handoff`. Inside routes/handoff.rs, `super::handoff` would
recurse to the module itself. The local `use crate::handoff;` makes
`handoff::MAX_BLOB_SIZE_BYTES` resolve to `crate::handoff::MAX_BLOB_SIZE_BYTES` —
the same constant, byte-identical runtime behavior.

All other fn bodies (spawn_routes, handoff_init_routes,
handoff_status_routes, 13 handlers) are byte-identical.

## Verification

- cargo test --package azureclaw-inference-router: 178/178
- cargo clippy --package azureclaw-inference-router --all-targets -- -D warnings: clean
- Drift: 1 intentional mutation (documented above)
